fix: caminhos inicial
