--- app/myecommerce/src/components/TopBar/index.tsx
@@ -1,5 +1,0 @@
-export default function TopBar() {
-    return (
-        <div>TOPBAR</div>
-    );
-}
